Added card custom component
