+++ Development/NutriMon/src/main/java/chillingMonsters/NutriMonMain.java
@@ -1,13 +1,6 @@
 package chillingMonsters;
 
-import java.util.List;
-import java.util.Map;
-
-import chillingMonsters.Controllers.ControllerFactory;
-import chillingMonsters.Controllers.IngredientController;
-import chillingMonsters.Controllers.NutriMonController;
-import chillingMonsters.Controllers.StockController;
-import chillingMonsters.Pages.loginPage;
+import chillingMonsters.Pages.loginPage.loginPage;
 import static javafx.application.Application.*;
 
 public class NutriMonMain {
