+++ Development/NutriMon/src/main/resources/stockPage/stock.fxml
@@ -5,22 +5,24 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
+<?import chillingMonsters.Pages.stockPage.*?>
 
-<AnchorPane fx:id="ancP_login_main" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="826.0" prefWidth="381.0" style="-fx-background-color: #FFFFFF;" styleClass="lebe;" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="NMUserProfile.loginController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="740.0" prefWidth="360.0" style="-fx-background-color: #FFFFFF;" styleClass="screenPane" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1">
    <children>
-      <ImageView fitHeight="33.0" fitWidth="33.0" layoutX="76.0" layoutY="113.0" pickOnBounds="true" preserveRatio="true" AnchorPane.leftAnchor="21.0" AnchorPane.topAnchor="50.0">
+      <ImageView fitHeight="33.0" fitWidth="33.0" layoutX="76.0" layoutY="113.0" pickOnBounds="true" preserveRatio="true" AnchorPane.leftAnchor="20.0" AnchorPane.topAnchor="45.0">
          <image>
-            <Image url="@img/Menu%20Icon2x.png" />
+            <Image url="@../img/Menu%20Icon2x.png" />
          </image>
       </ImageView>
-      <Label layoutX="121.0" layoutY="226.0" text="Your Stock" textFill="#f5a623" AnchorPane.leftAnchor="70.0" AnchorPane.topAnchor="52.5">
+      <Label layoutX="121.0" layoutY="226.0" text="Your Stock" textFill="#f5a623" AnchorPane.leftAnchor="65.0" AnchorPane.topAnchor="45.0">
          <font>
             <Font name="Helvetica" size="25.0" />
          </font></Label>
-      <Label layoutX="80.0" layoutY="63.0" text="Expires Soon" AnchorPane.leftAnchor="70.0" AnchorPane.topAnchor="120.0">
+      <Label text="Expires Soon" AnchorPane.leftAnchor="65.0" AnchorPane.topAnchor="100.0">
          <font>
             <Font name="Helvetica" size="20.0" />
          </font>
       </Label>
+      <StockCardComponent layoutX="30.0" layoutY="130.0" />
    </children>
 </AnchorPane>
