+++ Development/NutriMon/src/main/resources/stockPage/stockCard.fxml
@@ -1,54 +1,84 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.geometry.Insets?>
 <?import javafx.geometry.Rectangle2D?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.effect.DropShadow?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="95.0" prefWidth="315.0" style="-fx-background-color: #FFFFFFFF;" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1">
-  <children>
-     <Label text="Label" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="30.0">
-        <font>
-           <Font name="Helvetica" size="15.0" />
-        </font></Label>
-     <Label text="LABEL" textFill="#9e9e9e" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="10.0">
-        <font>
-           <Font name="Helvetica" size="9.0" />
-        </font>
-     </Label>
-     <ImageView fitHeight="15.0" fitWidth="15.0" layoutX="44.0" layoutY="7.0" pickOnBounds="true" preserveRatio="true" AnchorPane.bottomAnchor="15.0" AnchorPane.leftAnchor="15.0">
-        <image>
-           <Image url="@../img/ic_duration2x.png" />
-        </image>
-     </ImageView>
-     <Label layoutX="25.0" layoutY="40.0" text="Label" textFill="#9e9e9e" AnchorPane.bottomAnchor="15.0" AnchorPane.leftAnchor="40.0">
-        <font>
-           <Font name="Helvetica" size="12.0" />
-        </font>
-     </Label>
-     <ImageView fitHeight="15.0" fitWidth="15.0" layoutX="25.0" layoutY="70.0" pickOnBounds="true" preserveRatio="true" AnchorPane.bottomAnchor="15.0" AnchorPane.rightAnchor="185.0">
-        <image>
-           <Image url="@../img/ic_quantity2x.png" />
-        </image>
-     </ImageView>
-     <Label layoutX="50.0" layoutY="70.0" text="Label" textFill="#9e9e9e" AnchorPane.bottomAnchor="15.0" AnchorPane.rightAnchor="140.0">
-        <font>
-           <Font name="Helvetica" size="12.0" />
-        </font>
-     </Label>
-     <ImageView fitHeight="50.0" fitWidth="200.0" layoutX="130.0" layoutY="-45.0" pickOnBounds="true" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="-1.0">
-        <image>
-           <Image url="@../img/img_Kale.png" />
-        </image>
-        <viewport>
-           <Rectangle2D />
-        </viewport>
-     </ImageView>
-  </children>
-  <effect>
-     <DropShadow blurType="ONE_PASS_BOX" color="#c7c7c780" height="50.0" offsetX="10.0" offsetY="10.0" radius="37.0" width="100.0" />
-  </effect>
-</AnchorPane>
+<fx:root type="javafx.scene.layout.AnchorPane" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1">
+    <children>
+
+        <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="90.0" prefWidth="300.0" style="-fx-background-color: #FFFFFFFF;" styleClass="detailCardPane" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1">
+            <children>
+                <VBox AnchorPane.leftAnchor="10.0" AnchorPane.topAnchor="10.0">
+                    <children>
+                        <Label text="LABEL" textFill="#9e9e9e">
+                            <font>
+                                <Font name="Helvetica" size="9.0" />
+                            </font>
+                        </Label>
+                        <Label text="Label">
+                            <font>
+                                <Font name="Helvetica" size="15.0" />
+                            </font>
+                            <VBox.margin>
+                                <Insets top="10.0" />
+                            </VBox.margin>
+                        </Label>
+                    </children>
+                </VBox>
+                <HBox AnchorPane.leftAnchor="10.0" AnchorPane.topAnchor="60.0">
+                    <children>
+                        <ImageView fitHeight="15.0" fitWidth="15.0" pickOnBounds="true" preserveRatio="true">
+                            <image>
+                                <Image url="@../img/ic_duration2x.png" />
+                            </image>
+                        </ImageView>
+                        <Label text="Label" textFill="#9e9e9e">
+                            <font>
+                                <Font name="Helvetica" size="12.0" />
+                            </font>
+                            <HBox.margin>
+                                <Insets left="10.0" />
+                            </HBox.margin>
+                        </Label>
+                        <ImageView fitHeight="15.0" fitWidth="15.0" pickOnBounds="true" preserveRatio="true">
+                            <image>
+                                <Image url="@../img/ic_quantity2x.png" />
+                            </image>
+                            <HBox.margin>
+                                <Insets left="85.0" />
+                            </HBox.margin>
+                        </ImageView>
+                        <Label text="Label" textFill="#9e9e9e">
+                            <font>
+                                <Font name="Helvetica" size="12.0" />
+                            </font>
+                            <HBox.margin>
+                                <Insets left="10.0" />
+                            </HBox.margin>
+                        </Label>
+                    </children>
+                </HBox>
+                <ImageView fitHeight="50.0" fitWidth="150.0" layoutX="130.0" layoutY="-45.0" pickOnBounds="true" AnchorPane.rightAnchor="-1.0" AnchorPane.topAnchor="-1.0">
+                    <image>
+                        <Image url="@img/KaleCardBG.png" />
+                    </image>
+                    <viewport>
+                        <Rectangle2D />
+                    </viewport>
+                </ImageView>
+            </children>
+            <effect>
+                <DropShadow blurType="ONE_PASS_BOX" color="#c7c7c77f" height="100.0" offsetX="10.0" offsetY="10.0" radius="49.5" width="100.0" />
+            </effect>
+        </AnchorPane>
+
+    </children>
+</fx:root>
